Keeps sundays If you provide a vector of dates, this function will keep the sundays and the latest date

Keeps sundays If you provide a vector of dates, this function will keep the sundays and the latest date

keep_sundays(dates, format = "Uke isoweek_c-1/isoweek_c", keep_delete = TRUE)

Arguments

dates
Vector of dates

format
Choose between: "isoyearweek_c", "Uke isoweek_c", "isoyearweek_c-1/isoyearweek_c", "Uke isoweek_c-1/isoweek_c", "date"

keep_delete
Keep everything in the same format as provided

Examples

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "isoyearweek_c",
  keep_delete = TRUE
)

#>  [1] delete  delete  delete  delete  2020-01 delete  delete  delete  delete 
#> [10] delete  delete  2020-02 delete  delete  delete  delete  delete  delete 
#> [19] 2020-03 delete  delete  delete  delete  delete  delete  2020-04 delete 
#> [28] delete  delete  delete  delete  delete 
#> Levels: delete 2020-01 2020-02 2020-03 2020-04

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "Uke isoweek_c",
  keep_delete = TRUE
)

#>  [1] delete delete delete delete Uke 01 delete delete delete delete delete
#> [11] delete Uke 02 delete delete delete delete delete delete Uke 03 delete
#> [21] delete delete delete delete delete Uke 04 delete delete delete delete
#> [31] delete delete
#> Levels: delete Uke 01 Uke 02 Uke 03 Uke 04

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "isoyearweek_c-1/isoyearweek_c",
  keep_delete = TRUE
)

#>  [1] delete          delete          delete          delete         
#>  [5] 2019-52/2020-01 delete          delete          delete         
#>  [9] delete          delete          delete          2020-01/2020-02
#> [13] delete          delete          delete          delete         
#> [17] delete          delete          2020-02/2020-03 delete         
#> [21] delete          delete          delete          delete         
#> [25] delete          2020-03/2020-04 delete          delete         
#> [29] delete          delete          delete          delete         
#> 5 Levels: delete 2019-52/2020-01 2020-01/2020-02 ... 2020-03/2020-04

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "Uke isoweek_c-1/isoweek_c",
  keep_delete = TRUE
)

#>  [1] delete    delete    delete    delete    Uke 52/01 delete    delete   
#>  [8] delete    delete    delete    delete    Uke 01/02 delete    delete   
#> [15] delete    delete    delete    delete    Uke 02/03 delete    delete   
#> [22] delete    delete    delete    delete    Uke 03/04 delete    delete   
#> [29] delete    delete    delete    delete   
#> Levels: delete Uke 52/01 Uke 01/02 Uke 02/03 Uke 03/04

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "date",
  keep_delete = TRUE
)

#>  [1] delete     delete     delete     delete     2020-01-05 delete    
#>  [7] delete     delete     delete     delete     delete     2020-01-12
#> [13] delete     delete     delete     delete     delete     delete    
#> [19] 2020-01-19 delete     delete     delete     delete     delete    
#> [25] delete     2020-01-26 delete     delete     delete     delete    
#> [31] delete     delete    
#> Levels: delete 2020-01-05 2020-01-12 2020-01-19 2020-01-26

fhiplot::keep_sundays(
  dates = seq.Date(as.Date("2020-01-01"), as.Date("2020-02-01"), by=1),
  format = "date",
  keep_delete = FALSE
)

#> [1] "2020-01-05" "2020-01-12" "2020-01-19" "2020-01-26"